refactor: move function
Move isInternalPackage function into package-name module to be closer to other related functions
